+++ Exceptions.h
@@ -12,20 +12,20 @@
 
 - (MVNode *)createCFINode:(MVNode *)parent
                   caption:(NSString *)caption
-                 location:(uint32_t)location
-                   length:(uint32_t)length;
+                 location:(NSInteger)location
+                   length:(NSUInteger)length;
 
 - (MVNode *)createLSDANode:(MVNode *)parent
                    caption:(NSString *)caption
-                  location:(uint32_t)location
-                    length:(uint32_t)length
+                  location:(NSInteger)location
+                    length:(NSUInteger)length
             eh_frame_begin:(uint64_t)eh_frame_begin;
 
 - (MVNode *)createUnwindInfoHeaderNode:(MVNode *)parent
                                caption:(NSString *)caption
                               location:(uint32_t)location
                                 header:
-                                    (struct unwind_info_section_header const *)
-                                        unwind_info_section_header;
+    (struct unwind_info_section_header const *)
+    unwind_info_section_header;
 
 @end
