+++ FatLayout.mm
@@ -40,7 +40,7 @@ using namespace std;
 //----------------------------------------------------------------------------
 - (MVNode *)createHeaderNode:(MVNode *)parent
                      caption:(NSString *)caption
-                    location:(uint32_t)location
+                    location:(NSInteger)location
                   fat_header:(struct fat_header const *)fat_header {
     MVNodeSaver nodeSaver;
     MVNode *node = [parent
