+++ Layout.h
@@ -20,8 +20,8 @@
 @interface MVLayout : NSObject {
     MVNode *__weak rootNode;
     MVDataController *__weak dataController;
-    uint32_t imageOffset; // absolute physical offset of the image in binary
-    uint32_t imageSize;   // size of the image corresponds to this layout
+    NSInteger imageOffset; // absolute physical offset of the image in binary
+    NSUInteger imageSize;   // size of the image corresponds to this layout
     NSThread *backgroundThread;
     MVArchiver *archiver;
 }
@@ -31,7 +31,7 @@
 @property(nonatomic, readonly) MVArchiver *archiver;
 
 - (id)initWithDataController:(MVDataController *)dc rootNode:(MVNode *)node;
-- (void const *)imageAt:(uint32_t)location;
+- (void const *)imageAt:(NSInteger)location;
 - (void)printException:(NSException *)exception caption:(NSString *)caption;
 - (BOOL)is64bit;
 - (void)doMainTasks;
@@ -41,7 +41,7 @@
 
 - (MVNode *)createDataNode:(MVNode *)parent
                    caption:(NSString *)caption
-                  location:(uint32_t)location
-                    length:(uint32_t)length;
+                  location:(NSInteger)location
+                    length:(NSUInteger)length;
 
 @end
